Move github-related functions to separate file
